New start
Created a new folder called classes where all classes are inside. Use this folder to create new classes. The main class is the blockchain class which is used to create the blockchain and add all the blocks to the chain. The class iets in the main directory is left behind for later usage(if we gonna use it later)